Initial CSS styling
Creates the z_application_overrides css file and does some initial
styling so it is no longer just the black/white/gray Rails default.
Changes include adding the background image. Also made a minor change
to the menu bars so that the logout link is not smushed against the
user name.